Modified scripts and reorganized the folder structure
1. Added feature for user to efficiently build, install, or uninstall
   individual or multiple driver modules without relying on the
   qcom_drivers.sh script.
2. Updated README.
3. Renamed and organized driver module folder structure
