+++ src/linux/qcom_serial/qcom_serial.h
