+++ src/linux/qcom_serial/qtiDevInf.h
