+++ src/linux/qcom_usb/qtiDevInf.h
