+++ src/linux/qcom_usbnet/common.h
