+++ src/linux/qcom_usbnet/ipassignment.sh
