+++ src/linux/qcom_usbnet/qmap.c
