+++ src/linux/qcom_usbnet/qmi.h
